python/requirements-s7_python.txt
python-snap7==1.0.0
